stupidly didn't commit until now but these initial ones should be the version used for all initial temporal fixes

+++ GrowthConeAnalyzer/VISUALIZATION_TOOLKIT/GCAmakeFinalGrowthConeMask.m
@@ -10,8 +10,9 @@ function [ output_args ] = makeNeuriteMask( analInfo ,maskDir)
 %  end 
  % move the old files 
 %  copyfile(maskDir,maskDirBody); 
+if isdir(maskDir)
  rmdir(maskDir,'s'); 
-  
+end 
  mkdir(maskDir)
